+++ bimms/system/BIMMS.py
@@ -28,7 +28,7 @@ from time import sleep
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
 
-from .BIMMSHadware import BIMMSHardware
+from .BIMMSconfig import BIMMSconfig
 from ..utils import constants as cst
 
 ### for debug
@@ -40,60 +40,9 @@ verbose = True
 ##############################
 ## CLASS FOR BIMMS HANDLING ##
 ##############################
-class BIMMS(BIMMSHardware):
+class BIMMS(BIMMSconfig):
     def __init__(self, bimms_id=None, serialnumber=None):
-        super().__init__()
-
-        # To maintain connection use keep_on
-        self.switch_off = True
-        self.selected = False
-        self.interface_on = False
-
-        if isinstance(bimms_id, int):
-            if bimms_id in cst.BimmsSerialNumbers:
-                self.serialnumber = cst.BimmsSerialNumbers[bimms_id]
-                self.selected = True
-            else:
-                print(
-                    "warning 'bimms_id' not referentced: first device will be selected"
-                )
-                exit()
-        elif isinstance(serialnumber, str):
-            if serialnumber in cst.BimmsSerialNumbers.values():
-                self.serialnumber
-                self.selected = True
-            else:
-                print(
-                    "warning 'serialnumber' not referentced: first device will be selected"
-                )
-                exit()
-
-        if self.selected:
-            self.interface = ai.Andi(self.serialnumber)
-        else:
-            self.interface = ai.Andi()
-            self.serialnumber = self.interface.serialnumber
-        self.interface_on = True
-        if verbose:
-            print("device opened")
-        self.ID = 0
-
-        self.SPI_init_STM32()
-
-        self.ID = self.get_board_ID()
-
-        if self.ID == 0 or self.ID > 16:  # only 8 bimms for now
-            self.close()
-            if verbose:
-                raise ValueError(
-                    "Failed to communicate with STM32 MCU. Make sure that BIMMS is powered (try to reconnect USB)."
-                )
-            quit()
-
-        if verbose:
-            print("You are connected to BIMMS " + str(self.ID))
-
-        self.DIO_init()
+        super().__init__(bimms_id=bimms_id, serialnumber=serialnumber)
 
         # Try to load DC calibration file
         if self.Load_DCCal():
@@ -123,22 +72,6 @@ class BIMMS(BIMMSHardware):
                 print("Consider running Open-Short-Load calibration script.")
             self.OSLCalibration = False
 
-    def __del__(self):
-        if self.switch_off and self.interface_on:
-            self.close()
-
-    def close(self):
-        self.set_state(cst.STM32_stopped)
-        self.interface.close()
-        self.interface_on = False
-        if verbose:
-            print("device closed")
-
-    def keep_on(self):
-        self.switch_off = False
-
-    def keep_off(self):
-        self.switch_off = True
 
     def Load_DCCal(self):
         if not os.path.exists(self.cal_folder):
@@ -322,497 +255,6 @@ class BIMMS(BIMMSHardware):
         outfile.close()
         self.OSLCalibration = True
         self.Load_OSLCal()
-
-    ##############################################
-    ## AD2 Digital IO methods for gains control ##
-    ##############################################
-    def DIO_init(self):
-        self.interface.configure_digitalIO()
-        self.set_DIO_mode()
-
-    def set_DIO_mode(self):
-        IO_vector = 0
-        IO_vector += self.IO6_IO * cst.IO6
-
-        # IO_vector += self.IO7_IO * cst.IO7
-        # LEDs and IA gain IOs are always set as outputs
-        IO_vector += cst.LED_status
-        IO_vector += cst.LED_err
-        IO_vector += cst.CH1_A0_0
-        IO_vector += cst.CH1_A1_0
-        IO_vector += cst.CH1_A0_1
-        IO_vector += cst.CH1_A1_1
-        IO_vector += cst.CH2_A0_0
-        IO_vector += cst.CH2_A1_0
-        IO_vector += cst.CH2_A0_1
-        IO_vector += cst.CH2_A1_1
-        self.interface.digitalIO_set_as_output(IO_vector)
-
-    def set_DIO_output(self):
-        OUTPUT_vector = 0
-        OUTPUT_vector += self.IO6_value * cst.IO6
-        OUTPUT_vector += self.IO7_value * cst.IO7
-        # LEDs and IA gain IOs are always set as outputs
-        OUTPUT_vector += self.LED_status * cst.LED_status
-        OUTPUT_vector += self.LED_err * cst.LED_err
-        OUTPUT_vector += self.CH1_A0_0 * cst.CH1_A0_0
-        OUTPUT_vector += self.CH1_A1_0 * cst.CH1_A1_0
-        OUTPUT_vector += self.CH1_A0_1 * cst.CH1_A0_1
-        OUTPUT_vector += self.CH1_A1_1 * cst.CH1_A1_1
-        OUTPUT_vector += self.CH2_A0_0 * cst.CH2_A0_0
-        OUTPUT_vector += self.CH2_A1_0 * cst.CH2_A1_0
-        OUTPUT_vector += self.CH2_A0_1 * cst.CH2_A0_1
-        OUTPUT_vector += self.CH2_A1_1 * cst.CH2_A1_1
-
-        self.interface.digitalIO_output(OUTPUT_vector)
-
-    def set_LED_status(self, value=True):
-        if value:
-            self.LED_status = 1
-        else:
-            self.LED_status = 0
-        self.set_DIO_output()
-
-    def set_LED_error(self, value=True):
-        if value:
-            self.LED_err = 1
-        else:
-            self.LED_err = 0
-        self.set_DIO_output()
-
-    def set_gain_ch1_1(self, value):
-        if (
-            (value != 1) and (value != 2) and (value != 5) and (value != 10)
-        ):  # Invalid gain value
-            self.CH1_A0_0 = 0  # Gain is one
-            self.CH1_A1_0 = 0
-        if value == 1:
-            self.CH1_A0_0 = 0
-            self.CH1_A1_0 = 0
-        if value == 2:
-            self.CH1_A0_0 = 1
-            self.CH1_A1_0 = 0
-        if value == 5:
-            self.CH1_A0_0 = 0
-            self.CH1_A1_0 = 1
-        if value == 10:
-            self.CH1_A0_0 = 1
-            self.CH1_A1_0 = 1
-        self.set_DIO_output()
-
-    def set_gain_ch1_2(self, value):
-        if (
-            (value != 1) and (value != 2) and (value != 5) and (value != 10)
-        ):  # Invalid gain value
-            self.CH1_A0_1 = 0
-            self.CH1_A1_1 = 0  # Gain is one
-        if value == 1:
-            self.CH1_A0_1 = 0
-            self.CH1_A1_1 = 0
-        if value == 2:
-            self.CH1_A0_1 = 1
-            self.CH1_A1_1 = 0
-        if value == 5:
-            self.CH1_A0_1 = 0
-            self.CH1_A1_1 = 1
-        if value == 10:
-            self.CH1_A0_1 = 1
-            self.CH1_A1_1 = 1
-        self.set_DIO_output()
-
-    def set_gain_ch2_1(self, value):
-        if (
-            (value != 1) and (value != 2) and (value != 5) and (value != 10)
-        ):  # Invalid gain value
-            self.CH2_A0_0 = 0
-            self.CH2_A1_0 = 0  # Gain is one
-        if value == 1:
-            self.CH2_A0_0 = 0
-            self.CH2_A1_0 = 0
-        if value == 2:
-            self.CH2_A0_0 = 1
-            self.CH2_A1_0 = 0
-        if value == 5:
-            self.CH2_A0_0 = 0
-            self.CH2_A1_0 = 1
-        if value == 10:
-            self.CH2_A0_0 = 1
-            self.CH2_A1_0 = 1
-        self.set_DIO_output()
-
-    def set_gain_ch2_2(self, value):
-        if (
-            (value != 1) and (value != 2) and (value != 5) and (value != 10)
-        ):  # Invalid gain value
-            self.CH2_A0_1 = 0
-            self.CH2_A1_1 = 0  # Gain is one
-        if value == 1:
-            self.CH2_A0_1 = 0
-            self.CH2_A1_1 = 0
-        if value == 2:
-            self.CH2_A0_1 = 1
-            self.CH2_A1_1 = 0
-        if value == 5:
-            self.CH2_A0_1 = 0
-            self.CH2_A1_1 = 1
-        if value == 10:
-            self.CH2_A0_1 = 1
-            self.CH2_A1_1 = 1
-        self.set_DIO_output()
-
-    def set_gain_IA(self, channel=1, gain=1):
-        gain_array = np.array([1, 2, 4, 5, 10, 20, 25, 50, 100])
-        gain_IA1 = np.array([1, 2, 2, 5, 5, 10, 5, 10, 10])
-        gain_IA2 = np.array([1, 1, 2, 1, 2, 2, 5, 5, 10])
-        idx_gain = np.where(gain_array == gain)
-        idx_gain = idx_gain[0]
-        if idx_gain != None:
-            if channel == 1:
-                self.set_gain_ch1_1(gain_IA1[idx_gain])
-                self.set_gain_ch1_2(gain_IA2[idx_gain])
-            if channel == 2:
-                self.set_gain_ch2_1(gain_IA1[idx_gain])
-                self.set_gain_ch2_2(gain_IA2[idx_gain])
-        else:
-            if verbose:
-                print("WARNING: Wrong IA gain value. IA gain set to 1.")
-            if channel == 1:
-                self.set_gain_ch1_1(gain_IA1[0])
-                self.set_gain_ch1_2(gain_IA2[0])
-            if channel == 2:
-                self.set_gain_ch2_1(gain_IA1[0])
-                self.set_gain_ch2_2(gain_IA2[0])
-
-    #################################
-    ## STM32 communitation methods ##
-    #################################
-    def set_state(self, state):
-        """
-        Set the state of STM32
-
-        Parameters
-        ----------
-        state : int
-            either STM32_stopped, STM32_idle, STM32_locked, STM32_error = 0x03
-            defined in BIMMS_constants
-        """
-        value = cst.cmd_shift * cst.set_STM32_state + state
-        self.tx_2_STM32(value)
-
-    def set_STM32_stopped(self):
-        self.set_state(cst.STM32_stopped)
-
-    def set_STM32_idle(self):
-        self.set_state(cst.STM32_idle)
-
-    def set_STM32_locked(self):
-        self.set_state(cst.STM32_locked)
-
-    def set_STM32_error(self):
-        self.set_state(cst.STM32_error)
-
-    def get_state(self):
-        """
-        Get the state of the STM32
-
-        Returns
-        -------
-        state	: int
-            0: STM32_stopped
-            1: STM32_idle
-            2: STM32_locked
-            3: STM32_error
-        """
-        state = self.read_STM32_register(cst.state_add)
-        return state
-
-    def get_STM32_error(self):
-        error = self.read_STM32_register(cst.error_add)
-        return error
-
-    ################################
-    ## BIMMS measurements methods ##
-    ################################
-
-    def set_2_points_config(self):
-        self.StimNeg2VNeg = 1
-        self.StimPos2VPos = 1
-
-    def set_3_points_config(self):
-        pass
-
-    def set_4_points_config(self):
-        self.StimNeg2VNeg = 0
-        self.StimPos2VPos = 0
-
-    def set_current_excitation(
-        self,
-        coupling="DC",
-        differential_stim=True,
-        DC_feedback=False,
-        Internal_AWG=True,
-        High_gain=False,
-    ):
-        if coupling == "DC":
-            self.set_Stim_DC_coupling()
-            self.disable_DC_feedback()
-        else:
-            self.set_Stim_AC_coupling()
-        if DC_feedback:
-            self.enable_DC_feedback()
-        else:
-            self.disable_DC_feedback()
-        self.connect_Ipos_to_StimPos()
-        if differential_stim:
-            self.connect_Ineg_to_StimNeg()
-        else:
-            self.connect_GND_to_StimNeg()
-
-        if High_gain:
-            self.set_high_gain_current_source()
-        else:
-            self.set_low_gain_current_source()
-
-        if Internal_AWG:
-            self.connect_internal_AWG()
-        else:
-            self.connect_external_AWG()
-        self.disable_potentiostat()
-
-    def set_voltage_excitation(
-        self, coupling="DC", differential_stim=True, Internal_AWG=True
-    ):
-        self.disable_DC_feedback()
-        if coupling == "DC":
-            self.set_Stim_DC_coupling()
-        else:
-            self.set_Stim_AC_coupling()
-        self.connect_Vpos_to_StimPos()
-
-        if differential_stim:
-            self.connect_Vneg_to_StimNeg()
-        else:
-            self.connect_GND_to_StimNeg()
-
-        if Internal_AWG:
-            self.connect_internal_AWG()
-        else:
-            self.connect_external_AWG()
-        # self.disable_current_source()			#need to be tested, bug with AD830?
-        self.disable_potentiostat()
-
-    def set_recording_channel_1(self, coupling="DC", gain=1.0):
-        self.connect_CH1_to_scope_1()
-        if coupling == "DC":
-            self.set_CH1_DC_coupling()
-        else:
-            self.set_CH1_AC_coupling()
-
-        self.set_gain_IA(channel=1, gain=gain)
-
-    def set_recording_channel_2(self, coupling="DC", gain=1.0):
-        self.connect_CH2_to_scope_2()
-        if coupling == "DC":
-            self.set_CH2_DC_coupling()
-        else:
-            self.set_CH2_AC_coupling()
-
-        self.set_gain_IA(channel=2, gain=gain)
-
-    def set_recording_voltage(self, coupling="DC", gain=1.0):
-        self.set_recording_channel_1(coupling=coupling, gain=gain)
-
-    def set_recording_current(self, differential=True, coupling="DC", gain=1.0):
-        self.set_recording_channel_2(coupling=coupling, gain=gain)
-        self.connect_TIA_to_CH2()
-        self.connect_TIA_to_StimNeg()
-        # self.connect_TIA_Neg_to_ground()
-
-        if differential:
-            if self.VoutPos2StimPos:  # Voltage excitation
-                self.connect_TIA_Neg_to_Vneg()
-            else:
-                if self.connect_Ipos_to_StimPos():  # Current excitation
-                    self.connect_TIA_Neg_to_Ineg()
-                else:
-                    self.connect_TIA_Neg_to_ground()
-        else:
-            self.connect_TIA_Neg_to_ground()
-        if coupling == "DC":
-            self.set_TIA_DC_coupling()
-        else:
-            self.set_TIA_DC_coupling()
-
-    def set_potentiostat_EIS_config(
-        self,
-        differential=True,
-        two_wires=True,
-        coupling="DC",
-        voltage_gain=1,
-        current_gain=1,
-    ):
-        self.set_STM32_idle()
-        if differential:
-            if coupling == "DC":
-                self.set_voltage_excitation(coupling="DC", differential_stim=True)
-                self.set_recording_current(
-                    differential=True, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-            else:
-                self.set_voltage_excitation(coupling="AC", differential_stim=True)
-                self.set_recording_current(
-                    differential=True, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="AC", gain=voltage_gain)
-        else:
-            if coupling == "DC":
-                self.set_voltage_excitation(coupling="DC", differential_stim=False)
-                self.set_recording_current(
-                    differential=False, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-            else:
-                self.set_voltage_excitation(coupling="AC", differential_stim=False)
-                self.set_recording_current(
-                    differential=False, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="AC", gain=voltage_gain)
-        if two_wires:
-            self.set_2_points_config()
-        else:
-            self.set_4_points_config()
-        self.set_config()
-
-    def set_galvanostat_EIS_config(
-        self,
-        differential=True,
-        two_wires=True,
-        High_gain=True,
-        coupling="DC",
-        DC_feedback=False,
-        voltage_gain=1,
-        current_gain=1,
-    ):
-        self.set_STM32_idle()
-        if differential:
-            if coupling == "DC":
-                self.set_current_excitation(
-                    coupling="DC",
-                    differential_stim=True,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=True, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-            else:
-                self.set_current_excitation(
-                    coupling="AC",
-                    differential_stim=True,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=True, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-        else:
-            if coupling == "DC":
-                self.set_current_excitation(
-                    coupling="DC",
-                    differential_stim=False,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=False, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-            else:
-                self.set_current_excitation(
-                    coupling="AC",
-                    differential_stim=False,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=False, coupling="DC", gain=current_gain
-                )
-                self.set_recording_voltage(coupling="DC", gain=voltage_gain)
-        if two_wires:
-            self.set_2_points_config()
-        else:
-            self.set_4_points_config()
-
-        self.set_config()
-
-    def set_cyclic_voltametry_config(
-        self, mode="two_points", coupling="DC", differential=True
-    ):
-        self.set_STM32_idle()
-        if mode == "two_points":
-            if coupling == "DC":
-                self.set_voltage_excitation(
-                    coupling="DC", differential_stim=differential
-                )
-                self.set_recording_current(
-                    differential=differential, coupling="DC", gain=1
-                )
-                self.set_recording_voltage(coupling="DC", gain=1)
-            else:
-                self.set_voltage_excitation(
-                    coupling="AC", differential_stim=differential
-                )
-                self.set_recording_current(
-                    differential=differential, coupling="DC", gain=1
-                )
-                self.set_recording_voltage(coupling="AC", gain=1)
-            self.set_2_points_config()
-
-        self.set_config()
-
-    def set_cyclic_amperometry_config(
-        self,
-        mode="two_points",
-        coupling="DC",
-        differential=True,
-        High_gain=True,
-        DC_feedback=False,
-    ):
-        self.set_STM32_idle()
-        if mode == "two_points":
-            if coupling == "DC":
-                self.set_current_excitation(
-                    coupling="DC",
-                    differential_stim=True,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=differential, coupling="DC", gain=1
-                )
-                self.set_recording_voltage(coupling="DC", gain=1)
-            else:
-                self.set_current_excitation(
-                    coupling="DC",
-                    differential_stim=True,
-                    DC_feedback=DC_feedback,
-                    Internal_AWG=True,
-                    High_gain=High_gain,
-                )
-                self.set_recording_current(
-                    differential=differential, coupling="DC", gain=1
-                )
-                self.set_recording_voltage(coupling="AC", gain=1)
-            self.set_2_points_config()
-        self.set_config()
 
     #########################
     ## Calibration methods ##
