--- examples/CalibrationData/DCCal_BIMMS_4.json
@@ -1,1 +1,0 @@
-{"BIMMS_SERIAL": 4, "gain_voltage_SE": 1.0853861094587451, "gain_voltage_DIFF": 2.1719457203889667, "low_gain_current": 0.03165116558158011, "high_gain_current": 0.06904293381626514, "gain_TIA": 1004.6986397106465}
