+++ tests/software_unitary_tests/003_test_SPI.py
