+++ tests/software_unitary_tests/004_test_COM_STM32.py
